% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/attach_key.R
\name{attach_key}
\alias{attach_key}
\title{Attach a key}
\usage{
attach_key(map, mapkey)
}
\arguments{
\item{map}{An object generated with \code{\link{build_bmap}} or \code{\link{build_gmap}}.}

\item{mapkey}{An object generated with \code{\link{build_bkey}} or \code{\link{build_gkey}}.}
}
\description{
This function attaches a bivariate colour key to a
bivariate colour map or a glyph key to a glyph map so that they can be viewed together.
}
\examples{
data(us_data)
data(us_geo)
poverty <- read.uv(data = us_data, estimate = "pov_rate", error = "pov_moe")

# bivariate map and key together
map <- build_bmap(data = poverty, geoData = us_geo, id = "GEO_ID",
 border = "state", terciles = TRUE)
key <- build_bkey(data = poverty, terciles = TRUE)
attach_key(map, key)
}
