% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_bmap.R
\name{build_bmap}
\alias{build_bmap}
\title{Build a bivariate colour map}
\usage{
build_bmap(data, shapefile = NULL, id = NULL, border = NULL,
  palette = "BlueYellow", size = NULL, terciles = FALSE,
  bound = NULL, flipAxis = FALSE)
}
\arguments{
\item{data}{A data frame.}

\item{shapefile}{A spatial polygons data frame.}

\item{id}{Name of the common column shared by the objects passed to
\code{data} and \code{shapefile}. The estimates and errors in the data frame
will be matched to the geographical regions of the spatial polygons data
frame through this column.}

\item{border}{Name of geographical borders to be added to the map. It must be
one of \code{\link[maps]{county}}, \code{\link[maps]{france}},
\code{\link[maps]{italy}}, \code{\link[maps]{nz}},
\code{\link[maps]{state}}, \code{\link[maps]{usa}} or
\code{\link[maps]{world}} (see documentation for
\code{\link[ggplot2]{map_data}} for more information). The borders will be
refined to match latitute and longtidue coordinates provided in the data
frame or spatial polygons data frame.}

\item{palette}{Name of colour palette or character vector of hex colour codes
from the \code{\link{build_palette}} function. Colour palette names include
\code{BlueYellow}, \code{CyanMagenta}, \code{BlueRed} and \code{GreenBlue}.}

\item{size}{An integer between 1 and 20. Value controls the size of points
when \code{shapefile = NULL}. If \code{size = NULL}, the points will remain
the default size.}

\item{terciles}{A logical value. This provides the option to define numerical
bounds for the colour key grid using terciles instead of equal intervals.}

\item{bound}{Output from the \code{findNbounds} function if a different
set of data is required to bound the map.  This is useful if you are wanting
to create a bivariate map across multiple years and show colours that correspond
to the same key. Default is NULL.}

\item{flipAxis}{A logical value. Whether to place the axis on the opposite sides or not.}
}
\description{
This function builds a map that visualises estimates and errors simultaneously
with a bivariate colour scheme.
}
\details{
If \code{shapefile} remains \code{NULL}, the function will produce a map of
plotted points representing specific sites; in this case, the data frame must
include latitude and longitude coordinates in columns \code{"long"} and
\code{"lat"}.
}
\examples{
data(us_data)
data(us_geo)
poverty <- read.uv(data = us_data, estimate = "pov_rate", error = "pov_moe")

#bivariate map with a shapefile
map <- build_bmap(data = poverty, shapefile = us_geo, id = "GEO_ID",
 border = "state", terciles = TRUE)
view(map)

}
\seealso{
\code{\link{build_bkey}}
}
