% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_emap.r
\name{build_emap}
\alias{build_emap}
\title{Build an exceedance probability colour map}
\usage{
build_emap(data, shapefile = NULL, id = NULL, key_label,
  palette = "YlOrRd", size = NULL, border = NULL)
}
\arguments{
\item{data}{A data frame.}

\item{shapefile}{A spatial polygons data frame.}

\item{id}{Name of the common column shared by the objects passed to
\code{data} and \code{shapefile}. The exceedance probability in the data frame
will be matched to the geographical regions of the spatial polygons data
frame through this column.}

\item{key_label}{Label of legend.}

\item{palette}{Name of colour palette. Colour palette names include
\code{BlueYellow}, \code{CyanMagenta}, \code{BlueRed}, \code{GreenBlue} and \code{YellowRed}.}

\item{size}{An integer between 1 and 20. Value controls the size of points
when \code{shapefile = NULL}. If \code{size = NULL}, the points will remain
the default size.}

\item{border}{Name of geographical borders to be added to the map. It must be
one of \code{\link[maps]{county}}, \code{\link[maps]{france}},
\code{\link[maps]{italy}}, \code{\link[maps]{nz}},
\code{\link[maps]{state}}, \code{\link[maps]{usa}} or
\code{\link[maps]{world}} (see documentation for
\code{\link[ggplot2]{map_data}} for more information). The borders will be
refined to match latitute and longtidue coordinates provided in the data
frame or spatial polygons data frame.}
}
\description{
This function builds a map that visualises the probability of exceeding some
threshold of concern.
}
\details{
If \code{shapefile} remains \code{NULL}, the function will produce a map of
plotted points representing specific sites; in this case, the data frame must
include latitude and longitude coordinates in columns \code{"long"} and
\code{"lat"}.
}
