% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_gkey.R
\name{build_gkey}
\alias{build_gkey}
\title{Build a glyph key}
\usage{
build_gkey(data, glyph = "icone", transparent = FALSE)
}
\arguments{
\item{data}{A data frame.}

\item{glyph}{Name of glyph shape. Options include \code{icone} and
\code{semi}.}

\item{transparent}{A logical value. Option to make key background transparent.
Default is FALSE.}
}
\description{
This function creates a key of rotated glyphs for a map produced with
\code{\link{build_gmap}}.
}
\details{
A key for the glyph map is not automatically generated with
\code{\link{build_gmap}} and must be made using \code{\link{build_gkey}}. It
is important that the arguments passed to \code{\link{build_gkey}} match those
passed to \code{\link{build_gmap}}. The map and key can be viewed together
using \code{\link{attach_key}}.
}
\examples{
data(us_data)
data(us_geo)
co_geo <- subset(us_geo, us_geo@data$STATE == "08")
us_data$GEO.id2 <- as.numeric(us_data$GEO.id2)
co_data <- subset(us_data, us_data$GEO.id2 > 8000 & us_data$GEO.id2 < 9000)
co_data <- read.uv(data = co_data, estimate = "pov_rate", error = "pov_moe")

# build a glyph key
key <- build_gkey(data = co_data, glyph = "icone")
view(key)

}
\seealso{
\code{\link{attach_key}}
}
