% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_gmap.R
\name{build_gmap}
\alias{build_gmap}
\title{Build a glyph map}
\usage{
build_gmap(
  data,
  geoData = NULL,
  id = NULL,
  size = 50,
  border = NULL,
  glyph = "icone",
  palette = "Blues",
  limits = NULL,
  max_error = NULL
)
}
\arguments{
\item{data}{A data frame with estimates and errors,
formatted with \code{read.uv}.}

\item{geoData}{An sf or sp object.}

\item{id}{Name of the common column shared by the objects passed to
\code{data} and \code{geoData}. The estimates and errors in the data frame
will be matched to the geographical regions through this column.}

\item{size}{An integer between 1 and 100. Value controls the size of the glyphs.}

\item{border}{Name of geographical borders to be added to the map. It must be
one of \code{\link[maps]{county}}, \code{\link[maps]{france}},
\code{\link[maps]{italy}}, \code{\link[maps]{nz}},
\code{\link[maps]{state}}, \code{\link[maps]{usa}} or
\code{\link[maps]{world}} (see documentation for
\code{\link[ggplot2]{map_data}} for more information). The borders will be
refined to match latitute and longtidue coordinates from the geoData argument.}

\item{glyph}{Name of glyph shape. Options include \code{icone} and \code{semi}.}

\item{palette}{Name of colour palette. It must be one of \code{Blues},
\code{Greens}, \code{Greys}, \code{Oranges}, \code{Purples} or \code{Reds}
(see documentation for \code{\link[ggplot2]{scale_fill_distiller}} for more
information).}

\item{limits}{Limits for the legend. Default is NULL, which takes the limits
to be the range of the data.}

\item{max_error}{maximum error value. Default is NULL, which takes the
maximum from the error given}
}
\description{
This function builds a map that visualises estimates and errors simultaneously
with rotated glyphs.
}
\details{
If an sf or sp object is used, glyphs will be plotted at region
centroids. If \code{geoData} remains \code{NULL}, glyphs will be plotted at
points on the map representing specific sites; in this case, the data frame
must include latitude and longitude coordinates in columns \code{"long"} and
\code{"lat"}.
}
\examples{
data(us_data)
data(us_geo)
co_geo <- subset(us_geo, us_geo@data$STATE == "08")
us_data$GEO.id2 <- as.numeric(us_data$GEO.id2)
co_data <- subset(us_data, us_data$GEO.id2 > 8000 & us_data$GEO.id2 < 9000)
co_data <- read.uv(data = co_data, estimate = "pov_rate", error = "pov_moe")

# build a glyph map
map <- build_gmap(data = co_data, geoData = co_geo, id = "GEO_ID",
 size = 70, border = "state", glyph = "icone")
view(map)

}
\seealso{
\code{\link{build_gkey}}
}
