% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createN.r
\name{createN}
\alias{createN}
\title{Internal function to create random variables
Creates a Normal random variable}
\usage{
createN(x, est, error)
}
\arguments{
\item{x}{x object}

\item{est}{Estimate}

\item{error}{Error}
}
\description{
Internal function to create random variables
Creates a Normal random variable
}
