% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createPixrv.r
\name{createPixrv}
\alias{createPixrv}
\title{CreatePixrv}
\usage{
createPixrv(pixelGeo, distribution)
}
\arguments{
\item{pixelGeo}{A pixelGeo object}

\item{distribution}{Distribution. One of "uniform", "normal" or "discrete"}
}
\description{
This is an internal function.
}
