% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findNbounds.r
\name{findNbounds}
\alias{findNbounds}
\title{findNbounds}
\usage{
findNbounds(data, estimate, error, terciles)
}
\arguments{
\item{data}{dataset from which bounds are to be determined}

\item{estimate}{name of the estimate column}

\item{error}{name of the error column}

\item{terciles}{should terciles be calculated? (Default: FALSE)}
}
\description{
Find the bins for the colour grid
}
