% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdist.r
\name{pdist}
\alias{pdist}
\title{pdist}
\usage{
pdist(pname, th, args)
}
\arguments{
\item{pname}{Distribution to be used which is embedded through the \code{quote} command.
This can comprise an existing function from the \code{stats} package e.g. \code{pnorm}.
Alternatively this can be user defined.}

\item{th}{Threshold for calculating the \[Pr(X > th)\]}

\item{args}{Arguments that correspond to the specified distribution (but excluding the
threshold.}
}
\description{
Calculates the probability of exceedance for a given distribution,
threshold and distribution parameters.
}
\examples{
pname <- quote({ pnorm(q, mean, sd, lower.tail = FALSE) })
pdist(pname = pname, th = 3, args = list(mean = 2, sd = 1))


}
