% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pixelate.R
\name{pixelate}
\alias{pixelate}
\title{Pixelate polygons}
\usage{
pixelate(geoData = NULL, file = NULL, layer = NULL, pixelSize = 40, id = "id")
}
\arguments{
\item{geoData}{An object of class "SpatialPolygons" or
"SpatialPolygonsDataFrame".}

\item{file}{A shapefile pathway.}

\item{layer}{Name of geoData layer (see documentation for
\code{\link[sf]{read_sf}} for more information).}

\item{pixelSize}{An integer. Larger values result in smaller pixels.}

\item{id}{The ID column shared by the geoData object and the dataset with
estimates and errors.}
}
\description{
This function divides a shapefile into pixels so it can be used to create a
pixel map with \code{\link{build_pmap}}.
}
\details{
This function can take several minutes to run depending on the size of the shapefile. Within this function, the projection of the spatial object is removed and then returned to the original projection.
}
\examples{
data(us_geo)
ca_geo <- subset(us_geo, us_geo@data$STATE == "06")
pix <- pixelate(ca_geo, pixelSize = 60, id = "GEO_ID")

}
