% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read.uv.r
\name{read.uv}
\alias{read.uv}
\title{Format data}
\usage{
read.uv(data = NULL, file = NULL, estimate, error, exceedance)
}
\arguments{
\item{data}{A data frame.}

\item{file}{A CSV file pathway.}

\item{estimate}{Name of estimate column.}

\item{error}{Name of error column.}

\item{exceedance}{Name of exceedance probability column.}
}
\description{
This function formats a data frame so it can be used with
\code{\link{build_bmap}}, \code{\link{build_gmap}}, \code{\link{build_pmap}},
\code{\link{build_bkey}} and \code{\link{build_gkey}}.
}
\details{
Estimates and errors must be in the first and second columns of a data frame
for \code{\link{build_bmap}}, \code{\link{build_gmap}},
\code{\link{build_pmap}}, \code{\link{build_emap}},
\code{\link{build_bkey}} and \code{\link{build_gkey}}. \code{\link{read.uv}} provides an automated way to
format a data frame already loaded in R or to import a CSV file as a data
frame and arrange the columns correctly.
}
\examples{
data(us_data)
poverty <- read.uv(data = us_data, estimate = "pov_rate", error = "pov_moe")
head(poverty)

}
