% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UB_shp.r
\docType{data}
\name{UB_shp}
\alias{UB_shp}
\title{UB_shp}
\format{
A shape file stored as a \code{SpatialPolygonsDataFrame}.
}
\usage{
UB_shp
}
\description{
Upper Burdekin Shape file consisting of 411 sub-catchment
boundaries where estimates of Total Suspended Sediment (TSS) has been
quantified.
}
\keyword{datasets}
