% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view.r
\name{view}
\alias{view}
\title{View a \strong{VizU} object}
\usage{
view(obj)
}
\arguments{
\item{obj}{An object from \code{\link{build_bmap}}, \code{\link{build_gmap}},
\code{\link{build_pmap}}, \code{\link{build_bkey}},
\code{\link{build_gkey}}, \code{\link{build_palette}} or
\code{\link{animate}}.}
}
\description{
This function views maps, keys, palettes and animations created with
\strong{VizU}.
}
\details{
For maps, keys and palettes, assigning the output from \code{\link{view}} to a variable
name saves the output as an object of class \code{"ggplot"}. Animations
created with \code{\link{animate}} are viewed through an internet browser
as html objects.
}
\examples{
gb <- build_palette(name = "GreenBlue")
view(gb)

#ggplot object
p <- view(gb)


}
