\name{VizU-internal}
\alias{calcbin.r}
\alias{findNbounds.r}
\alias{view.bivmap.r}
\alias{view.key.r}
\alias{view.palette.r}



\title{Internal VizU functions}
\description{
These are internal \code{VizU} functions that are not intended to be directly called by the user
}


\author{Lydia Lucchesi and Petra Kuhnert}
\keyword{internal}
