\name{Vizumap-internal}
\alias{calcbin.r}
\alias{findNbounds.r}
\alias{view.bivmap.r}
\alias{view.key.r}
\alias{view.palette.r}



\title{Internal Vizumap functions}
\description{
These are internal \code{Vizumap} functions that are not intended to be directly called by the user
}


\author{Petra Kuhnert and Lydia Lucchesi}
\keyword{internal}
