% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_palette.r
\name{build_palette}
\alias{build_palette}
\title{Build a palette}
\usage{
build_palette(name, colrange = list(colour = NULL, difC = NULL),
  flipVertical = FALSE, flipHorizontal = FALSE)
}
\arguments{
\item{name}{Name of colour palette or \code{usr} for option to design a new
palette. Colour palette names include
\code{BlueYellow}, \code{CyanMagenta}, \code{BlueRed} and
\code{GreenBlue}.}

\item{colrange}{List with a character vector of length two and a numeric
vector of length two.}

\item{flipVertical}{Whether the palette should be flipped vertically (ie. replace top portion with bottom portion)}

\item{flipHorizontal}{Whether the palette should be flipped horizontally (ie. replace left portion with right portion)}

\item{colour}{A character vector of two colour names from the colours() range or valid hexadecimal colors.}

\item{difC}{A numeric vector of two integers 1, 2, 3 or 4. Values
control how much a colour changes in value across the grid. One corresponds
with a small change in colour value, and four corresponds with a large
change in colour value.}
}
\description{
This function prepares one of the four included colour palettes
or builds a new colour palette.
}
\details{
Note that \code{colrange} only needs to be specified if \code{name = "usr."}
When choosing colours, it is best to avoid light colours or tints as these
will lead to a colour palette lacking noticeable differences across the 3 x 3
colour grid.
}
\examples{
#use one of four prepared colour palettes
p <- build_palette(name = "CyanMagenta")
view(p)

#design a new palette
p <- build_palette(name = "usr", colrange =
 list(colour = c("darkblue", "chartreuse4"), difC = c(3, 4)))
view(p)
}
