% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pixelate.R
\name{pixelate}
\alias{pixelate}
\title{Pixelate polygons}
\usage{
pixelate(shapefile = NULL, file = NULL, layer = NULL, pixelSize = 2, id = "id")
}
\arguments{
\item{shapefile}{An object of class "SpatialPolygons" or
"SpatialPolygonsDataFrame".}

\item{file}{A shapefile pathway.}

\item{layer}{Name of shapefile layer (see documentation for
\code{\link[rgdal]{readOGR}} for more information).}

\item{pixelSize}{An integer 1, 2 or 3. One corresponds to the smallest pixel
size, and three corresponds to the largest.}

\item{id}{A name which will be given to the new ID column. This ID corresponds to the slot ID in
the shapefile.}
}
\description{
This function divides a shapefile into pixels so it can be used to create a
pixel map with \code{\link{build_pmap}}.
}
\details{
This function can take several minutes to run depending on the size of the shapefile.
}
\examples{
data(us_geo)
ca_geo <- subset(us_geo, us_geo@data$STATE == "06")
pix <- pixelate(ca_geo, id = "region")


}
